Move fluent Firestore classes into their own package
